+++ src/user-py/share/Makefile.common.targets
@@ -11,7 +11,6 @@ valgrind-check: $(TARGET)
 deploy:
 	test -d $(DEPLOY_DIR) || mkdir -p $(DEPLOY_DIR)
 	test -d $(COMMON_DIR) || mkdir -p $(COMMON_DIR)
-	touch $(COMMON_PARENT_DIR)/__init__.py
 	cp -rf ./ $(COMMON_DIR)
 
 clean:
