+++ src/user-py/share/Makefile.mod
@@ -3,7 +3,7 @@ TARGET := all
 MODULE_NAME ?= $(shell basename $(shell pwd))
 CONF_TARGET ?= skull-modules-$(MODULE_NAME).yaml
 
-MODULE_PARENT_DIR := $(DEPLOY_DIR)/lib/py/skull/modules
+MODULE_PARENT_DIR := $(DEPLOY_DIR)/lib/py/modules
 MODULE_DIR := $(MODULE_PARENT_DIR)/$(MODULE_NAME)
 
 # Include the basic Makefile targets
