+++ src/user-py/share/module.py
@@ -1,11 +1,11 @@
 import yaml
 import pprint
 
-from skullpy     import *
-from skullpy.txn import *
+from skull     import *
+from skull.txn import *
 
-from skull.common import *
-from skull.common.proto import *
+from common import *
+from common.proto import *
 
 ##
 # Module Init Entry, be called when start phase
