+++ src/user-py/share/module_http.py
@@ -2,14 +2,14 @@ import yaml
 import pprint
 import sys
 
-from skullpy import txn     as Txn
-from skullpy import txndata as TxnData
-from skullpy import logger  as Logger
-from skullpy import http
+from skull import txn     as Txn
+from skull import txndata as TxnData
+from skull import logger  as Logger
+from skull import http
 
-from skull.common import protos  as Protos
-from skull.common import metrics as Metrics
-from skull.common.proto import *
+from common import protos  as Protos
+from common import metrics as Metrics
+from common.proto import *
 
 ##
 # Module Init Entry, be called when start phase
